backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flask-migrate==4.0.5
google-generativeai==0.8.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
scikit-learn==1.4.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
